# docs/api.yml
openapi: 3.0.3
info:
  title: Red Anti-Social
  version: 1.0.1
  description: Documentación API de Red Anti-Social.
paths:
...
  /posts:
    get:
      summary: Obtener todos los posts
      tags:
        - Posts
      description: Recupera la lista de posts (desde Redis si está en caché). Oculta de forma automática comentarios con más de 6 meses de antigüedad y devuelve datos populados de autor, tags y comentarios visibles.
      responses:
        '200':
          description: Lista de posts obtenida correctamente.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'
        '500':
          description: Error al obtener los posts.

    post:
      summary: Crear un nuevo post
      tags:
        - Posts
      description: Crea un post e invalida la caché de Redis.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - texto
                - autor
              properties:
                texto:
                  type: string
                  example: "Hoy me levanté con menos ganas de ver gente que ayer."
                autor:
                  type: string
                  description: "ID del usuario creador"
                  example: "64f1bc23e4b0a123456789ab"
                imagenes:
                  type: array
                  items:
                    type: string
                  example: ["https://imagenes.com/meme1.jpg"]
                tags:
                  type: array
                  items:
                    type: string
                  example: ["64f1bc23e4b0a12345678902"]
      responses:
        '201':
          description: Post creado correctamente.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Post creado correctamente"
                  post:
                    $ref: '#/components/schemas/Post'
        '500':
          description: Error al crear el post.
...
  /comentarios:
    post:
      summary: Crear un nuevo comentario
      tags:
        - Comentarios
      description: Crea un comentario y añade automáticamente su ID al array de comentarios del post relacionado.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - contenido
                - autor
                - post
              properties:
                contenido:
                  type: string
                  example: "Totalmente de acuerdo con el desconocido."
                autor:
                  type: string
                  description: "ID del usuario que comenta"
                  example: "64f1bc23e4b0a123456789ab"
                post:
                  type: string
                  description: "ID del post a comentar"
                  example: "64f1bc23e4b0a123456789ef"
      responses:
        '201':
          description: Comentario creado y vinculado correctamente.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Comentario creado correctamente"
                  comentario:
                    $ref: '#/components/schemas/Comment'
        '500':
          description: Error al crear el comentario.
...
components:
  schemas:
...
    PostImage:
      type: object
      required:
        - url
      properties:
        _id:
          type: string
          example: "64f1bc23e4b0a123456789cd"
        url:
          type: string
          example: "https://imagenes.com/post1.jpg"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Post:
      type: object
      required:
        - texto
        - autor
      properties:
        _id:
          type: string
          example: "64f1bc23e4b0a123456789ef"
        texto:
          type: string
          example: "Este es un post verdaderamente anti-social."
        imagenes:
          type: array
          items:
            $ref: '#/components/schemas/PostImage'
        autor:
          type: string
          description: "ID del usuario creador"
          example: "64f1bc23e4b0a123456789ab"
        comentarios:
          type: array
          items:
            type: string
            description: "IDs de los comentarios asociados"
        tags:
          type: array
          items:
            type: string
            description: "IDs de los tags asociados"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Comment:
      type: object
      required:
        - contenido
        - autor
        - post
      properties:
        _id:
          type: string
          example: "64f1bc23e4b0a12345678901"
        contenido:
          type: string
          example: "No podría estar más en desacuerdo."
        autor:
          type: string
          description: "ID del usuario que comentó"
          example: "64f1bc23e4b0a123456789ab"
        post:
          type: string
          description: "ID del post comentado"
          example: "64f1bc23e4b0a123456789ef"
        visible:
          type: boolean
          default: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...